+++ src/lib/configParsing.js
@@ -27,6 +27,12 @@ export const KEY_VALUES = {
             id: "orientation", "canBeGlobal": true, value: "horizontal", type: "string",
             enum: ["vertical", "horizontal"],
             description: "The orientation of the graph. If you want to work from top to bottom or from bottom to top, set `orientation` to `vertical`"
+        },
+        "SPACING": {
+            id: "spacing", canBeGlobal: true, value: 1, type: "number",
+            description: "Set the spacing between nodes when autolayout, vgroups or hgroups are used." +
+                "The default spacing for groups is the width/2 for hgroups or the height/2 for vgroups devided by 2. If you want that larger or smaller, you can set this key to any number. The spacing will be the default times the factor." +
+                "When 'autolayout' is used, spacing has only effect on the vertical distance between nodes, when the orientation is set to 'vertical'. Vice versa for 'horizontal'."
         }
 
     },
@@ -421,7 +427,7 @@ export function parseNodes(globalDefaults, nodes) {
         }
 
     }
-    fixNodeGroups(nodes);
+    fixNodeGroups(globalDefaults, nodes);
 
     return nodes;
 }